add abstract pallet service
