+++ src/services/pallets/PalletsErrorsService.ts
@@ -15,25 +15,23 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import { ApiDecoration, ModuleErrors } from '@polkadot/api/types';
-import { Text, Vec } from '@polkadot/types';
+import { Text } from '@polkadot/types';
 import {
 	BlockHash,
 	ErrorMetadataLatest,
-	MetadataV14,
+	ModuleMetadataV13,
 	PalletErrorMetadataV14,
 	PalletMetadataV14,
 } from '@polkadot/types/interfaces';
 import { IsError } from '@polkadot/types/metadata/decorate/types';
 import { stringCamelCase } from '@polkadot/util';
-import { BadRequest, InternalServerError } from 'http-errors';
+import { InternalServerError } from 'http-errors';
 import {
 	IPalletErrors,
 	IPalletErrorsItem,
-	ISanitizedErrorItemMetadata,
 } from 'src/types/responses';
 
-import { sanitizeNumbers } from '../../sanitize/sanitizeNumbers';
-import { AbstractService } from '../AbstractService';
+import { PalletsService } from './PalletsService';
 
 interface IFetchPalletArgs {
 	hash: BlockHash;
@@ -45,17 +43,19 @@ interface IFetchErrorItemArgs extends IFetchPalletArgs {
 	metadata: boolean;
 }
 
-export class PalletsErrorService extends AbstractService {
+export class PalletsErrorService extends PalletsService {
 	async fetchErrorItem(
 		historicApi: ApiDecoration<'promise'>,
 		{ hash, palletId, errorItemId, metadata }: IFetchErrorItemArgs
 	): Promise<IPalletErrorsItem> {
+		const palletType = 'errors';
 		const palletMetadata = historicApi.registry.metadata;
 
 		const [palletMeta, palletMetaIdx] = this.findPalletMeta(
 			palletMetadata,
 			historicApi,
-			palletId
+			palletId,
+			palletType
 		);
 
 		const palletName = stringCamelCase(palletMeta.name);
@@ -69,18 +69,21 @@ export class PalletsErrorService extends AbstractService {
 			errors
 		);
 
-		let normalizedErrorItemMeta;
+		// let normalizedErrorItemMeta;
+		// if (metadata) {
+		// 	normalizedErrorItemMeta = this.normalizeErrorItemMeta(
+		// 		(errorItemMetadata[1] as IsError).meta
+		// 	);
+		// }
+		let palletErrorMetadata; 
 		if (metadata) {
-			normalizedErrorItemMeta = this.normalizeErrorItemMeta(
-				(errorItemMetadata[1] as unknown as IsError).meta
-			);
+			palletErrorMetadata = (errorItemMetadata[1] as IsError).meta
 		}
 
-		const [, { number }] = await Promise.all([
-			historicApi.query[palletName][errorItemId],
-			this.api.rpc.chain.getHeader(hash),
-		]);
+		console.log("pallet error metadata---", palletErrorMetadata);
 
+		const { number } = await this.api.rpc.chain.getHeader(hash);
+	
 		return {
 			at: {
 				hash: hash,
@@ -89,7 +92,7 @@ export class PalletsErrorService extends AbstractService {
 			pallet: stringCamelCase(palletMeta.name),
 			palletIndex: palletMetaIdx,
 			errorItem: errorItemId,
-			metadata: normalizedErrorItemMeta,
+			metadata: palletErrorMetadata
 		};
 	}
 
@@ -97,18 +100,20 @@ export class PalletsErrorService extends AbstractService {
 		historicApi: ApiDecoration<'promise'>,
 		{ hash, palletId, onlyIds }: IFetchPalletArgs & { onlyIds: boolean }
 	): Promise<IPalletErrors> {
+		const palletType = 'errors';
 		const metadata = historicApi.registry.metadata;
 		const [palletMeta, palletMetaIdx] = this.findPalletMeta(
 			metadata,
 			historicApi,
-			palletId
+			palletId,
+			palletType
 		);
 
 		const { number } = await this.api.rpc.chain.getHeader(hash);
 		const errors = historicApi.errors[palletMeta.name.toString().toLowerCase()];
 
 		let items: [] | ErrorMetadataLatest[] | Text[];
-		if (palletMeta.errors.isNone) {
+		if ((palletMeta.errors as unknown as PalletErrorMetadataV14).isEmpty) {
 			items = [];
 		} else if (onlyIds) {
 			items = Object.entries(errors).map(
@@ -134,36 +139,17 @@ export class PalletsErrorService extends AbstractService {
 	}
 
 	/**
-	 * Normalize error item metadata by running it through `sanitizeNumbers` and
-	 * converting the docs section from an array of strings to a single string
-	 * joined with new line characters.
-	 *
-	 * @param errorItemMeta polkadot-js ErrorEntryMetadataV12
-	 */
-	private normalizeErrorItemMeta(
-		errorItemMeta: ErrorMetadataLatest
-	): ISanitizedErrorItemMetadata {
-		const normalizedErrorItemMeta = sanitizeNumbers(
-			errorItemMeta
-		) as unknown as ISanitizedErrorItemMetadata;
-
-		normalizedErrorItemMeta.docs = this.sanitizeDocs(errorItemMeta.docs);
-
-		return normalizedErrorItemMeta;
-	}
-
-	/**
 	 * Find the error item's metadata within the pallets's metadata.
 	 *
 	 * @param palletMeta the metadata of the pallet that contains the error item
 	 * @param errorId name of the error item in camel or pascal case
 	 */
 	private findErrorItemMeta(
-		palletMeta: PalletMetadataV14,
+		palletMeta: PalletMetadataV14 | ModuleMetadataV13,
 		errorItemId: string,
 		errors: ModuleErrors<'promise'>
 	): PalletErrorMetadataV14 {
-		if (palletMeta.errors.isNone) {
+		if ((palletMeta.errors as unknown as PalletErrorMetadataV14).isEmpty) {
 			throw new InternalServerError(
 				`No error items found in ${palletMeta.name.toString()}'s metadata`
 			);
@@ -185,114 +171,5 @@ export class PalletsErrorService extends AbstractService {
 		return Object.entries(errors)[
 			errorItemMetaIdx
 		] as unknown as PalletErrorMetadataV14;
-	}
-
-	/**
-	 * Find a pallet's metadata info.
-	 *
-	 * @param palletId identifier for a FRAME pallet as a pallet name or index.
-	 */
-	private findPalletMeta(
-		adjustedMetadata: MetadataV14,
-		historicApi: ApiDecoration<'promise'>,
-		palletId: string
-	): [PalletMetadataV14, number] {
-		const pallets: Vec<PalletMetadataV14> = adjustedMetadata['pallets'];
-		const filtered: PalletMetadataV14[] = pallets.filter(
-			(mod) => mod.errors.isSome
-		);
-
-		const { isValidPalletName, isValidPalletIndex, parsedPalletId } =
-			this.validPalletId(historicApi, pallets, palletId);
-
-		let palletMeta: PalletMetadataV14 | undefined;
-		let palletIdx: number | undefined;
-
-		if (isValidPalletIndex) {
-			palletIdx = parsedPalletId as number;
-			for (const [_sectionIdx, section] of filtered.entries()) {
-				const idx = section.index.eqn(255)
-					? _sectionIdx
-					: section.index.toNumber();
-
-				if (idx === palletIdx) {
-					palletMeta = section;
-					break;
-				}
-			}
-		} else if (isValidPalletName) {
-			for (const [_sectionIdx, section] of filtered.entries()) {
-				if (section.name.toLowerCase() === palletId.toLowerCase()) {
-					// ModuleMetadataV11 and lower have an `index` but they use 255 as a reserve value to signify
-					// that they are meaningless. So if the index is 255 we use its index in the filtered array
-					// of modules. But if the index is something else than we use `ModuleMetadataV12.index`.
-					// The reason they use a reserve value is that all previous ModuleMetadata versions actually
-					// extend the latest. So since the intro of ModuleMetadataV12 all versions have `index` in
-					// polkadot-js, but at the substrate level, only versions >= 12 have pallet `index`.
-					// https://github.com/polkadot-js/api/pull/2599
-					// https://github.com/paritytech/substrate/pull/6969
-					// https://github.com/polkadot-js/api/issues/2596
-					palletIdx = section.index.eqn(255)
-						? _sectionIdx
-						: section.index.toNumber();
-					palletMeta = section;
-					break;
-				}
-			}
-		}
-
-		if (!palletMeta || palletIdx === undefined || palletIdx < 0) {
-			throw new BadRequest(
-				`"${palletId}" was not recognized as a queryable pallet.`
-			);
-		}
-
-		return [palletMeta, palletIdx];
-	}
-
-	private validPalletId(
-		historicApi: ApiDecoration<'promise'>,
-		modules: Vec<PalletMetadataV14>,
-		palletId: string
-	): {
-		isValidPalletName: boolean;
-		isValidPalletIndex: boolean;
-		parsedPalletId: string | number;
-	} {
-		// Either a pallet name (string) or a pallet index (number)
-		const parsedPalletId = PalletsErrorService.palletIdxOrName(palletId);
-
-		const isValidPalletName =
-			typeof parsedPalletId === 'string' && !!historicApi.query[palletId];
-
-		const isValidPalletIndex =
-			typeof parsedPalletId === 'number' &&
-			modules.some((meta, idx) =>
-				meta.index.eqn(255)
-					? idx === parsedPalletId
-					: meta.index.eqn(parsedPalletId)
-			);
-
-		return {
-			isValidPalletName,
-			isValidPalletIndex,
-			parsedPalletId,
-		};
-	}
-
-	/**
-	 * Identify if a pallet Identifier should be an index or a string. If it should
-	 * be an index return a number and if it should be a name return a string.
-	 *
-	 * @param palletId FRAME pallet identifier as a pallet name or index
-	 */
-	private static palletIdxOrName(palletId: string): string | number {
-		const maybeIdx = Number(palletId);
-
-		if (Number.isInteger(maybeIdx)) {
-			return maybeIdx;
-		}
-
-		return palletId;
 	}
 }
