+++ src/services/pallets/PalletsStorageService.ts
@@ -15,7 +15,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import { ApiDecoration } from '@polkadot/api/types';
-import { Text, Vec } from '@polkadot/types';
+import { Text } from '@polkadot/types';
 import {
 	BlockHash,
 	MetadataV13,
@@ -26,7 +26,7 @@ import {
 	StorageEntryMetadataV14,
 } from '@polkadot/types/interfaces';
 import { stringCamelCase } from '@polkadot/util';
-import { BadRequest, InternalServerError } from 'http-errors';
+import { InternalServerError } from 'http-errors';
 import {
 	IPalletStorage,
 	IPalletStorageItem,
@@ -34,7 +34,7 @@ import {
 } from 'src/types/responses';
 
 import { sanitizeNumbers } from '../../sanitize/sanitizeNumbers';
-import { AbstractService } from '../AbstractService';
+import { PalletsService } from './PalletsService';
 
 interface IFetchPalletArgs {
 	hash: BlockHash;
@@ -58,7 +58,7 @@ const upgradeBlocks = {
 	westend: 7766392,
 };
 
-export class PalletsStorageService extends AbstractService {
+export class PalletsStorageService extends PalletsService {
 	async fetchStorageItem(
 		historicApi: ApiDecoration<'promise'>,
 		{
@@ -70,6 +70,7 @@ export class PalletsStorageService extends AbstractService {
 			adjustMetadataV13Arg,
 		}: IFetchStorageItemArgs
 	): Promise<IPalletStorageItem> {
+		const palletType = 'storage';
 		const chosenMetadata = await this.chooseMetadataVersion(
 			historicApi,
 			hash,
@@ -78,7 +79,8 @@ export class PalletsStorageService extends AbstractService {
 		const [palletMeta, palletMetaIdx] = this.findPalletMeta(
 			chosenMetadata,
 			historicApi,
-			palletId
+			palletId,
+			palletType
 		);
 		const palletName = stringCamelCase(palletMeta.name);
 
@@ -120,6 +122,7 @@ export class PalletsStorageService extends AbstractService {
 			adjustMetadataV13Arg,
 		}: IFetchPalletArgs & { onlyIds: boolean }
 	): Promise<IPalletStorage> {
+		const palletType = 'storage';
 		const chosenMetadata = await this.chooseMetadataVersion(
 			historicApi,
 			hash,
@@ -128,7 +131,8 @@ export class PalletsStorageService extends AbstractService {
 		const [palletMeta, palletMetaIdx] = this.findPalletMeta(
 			chosenMetadata,
 			historicApi,
-			palletId
+			palletId,
+			palletType
 		);
 
 		let items: [] | ISanitizedStorageItemMetadata[] | Text[];
@@ -238,121 +242,5 @@ export class PalletsStorageService extends AbstractService {
 		}
 
 		return palletMetaStorage[storageItemMetaIdx];
-	}
-
-	/**
-	 * Find a pallet's metadata info.
-	 *
-	 * @param palletId identifier for a FRAME pallet as a pallet name or index.
-	 */
-	private findPalletMeta(
-		adjustedMetadata: MetadataV13 | MetadataV14,
-		historicApi: ApiDecoration<'promise'>,
-		palletId: string
-	): [PalletMetadataV14 | ModuleMetadataV13, number] {
-		const metadataType = adjustedMetadata.toRawType();
-
-		let pallets: Vec<PalletMetadataV14> | Vec<ModuleMetadataV13>;
-		let filtered: PalletMetadataV14[] | ModuleMetadataV13[];
-		if (metadataType.includes('MetadataV13')) {
-			pallets = adjustedMetadata['modules'] as Vec<ModuleMetadataV13>;
-			filtered = pallets.filter((mod) => mod.storage.isSome);
-		} else {
-			pallets = adjustedMetadata['pallets'] as Vec<PalletMetadataV14>;
-			filtered = pallets.filter((mod) => mod.storage.isSome);
-		}
-
-		const { isValidPalletName, isValidPalletIndex, parsedPalletId } =
-			this.validPalletId(historicApi, pallets, palletId);
-
-		let palletMeta: PalletMetadataV14 | ModuleMetadataV13 | undefined;
-		let palletIdx: number | undefined;
-
-		if (isValidPalletIndex) {
-			palletIdx = parsedPalletId as number;
-			for (const [_sectionIdx, section] of filtered.entries()) {
-				const idx = section.index.eqn(255)
-					? _sectionIdx
-					: section.index.toNumber();
-
-				if (idx === palletIdx) {
-					palletMeta = section;
-					break;
-				}
-			}
-		} else if (isValidPalletName) {
-			for (const [_sectionIdx, section] of filtered.entries()) {
-				if (section.name.toLowerCase() === palletId.toLowerCase()) {
-					// ModuleMetadataV11 and lower have an `index` but they use 255 as a reserve value to signify
-					// that they are meaningless. So if the index is 255 we use its index in the filtered array
-					// of modules. But if the index is something else than we use `ModuleMetadataV12.index`.
-					// The reason they use a reserve value is that all previous ModuleMetadata versions actually
-					// extend the latest. So since the intro of ModuleMetadataV12 all versions have `index` in
-					// polkadot-js, but at the substrate level, only versions >= 12 have pallet `index`.
-					// https://github.com/polkadot-js/api/pull/2599
-					// https://github.com/paritytech/substrate/pull/6969
-					// https://github.com/polkadot-js/api/issues/2596
-					palletIdx = section.index.eqn(255)
-						? _sectionIdx
-						: section.index.toNumber();
-					palletMeta = section;
-					break;
-				}
-			}
-		}
-
-		if (!palletMeta || palletIdx === undefined || palletIdx < 0) {
-			throw new BadRequest(
-				`"${palletId}" was not recognized as a queryable pallet.`
-			);
-		}
-
-		return [palletMeta, palletIdx];
-	}
-
-	private validPalletId(
-		historicApi: ApiDecoration<'promise'>,
-		modules: Vec<PalletMetadataV14> | Vec<ModuleMetadataV13>,
-		palletId: string
-	): {
-		isValidPalletName: boolean;
-		isValidPalletIndex: boolean;
-		parsedPalletId: string | number;
-	} {
-		// Either a pallet name (string) or a pallet index (number)
-		const parsedPalletId = PalletsStorageService.palletIdxOrName(palletId);
-
-		const isValidPalletName =
-			typeof parsedPalletId === 'string' && !!historicApi.query[palletId];
-
-		const isValidPalletIndex =
-			typeof parsedPalletId === 'number' &&
-			modules.some((meta, idx) =>
-				meta.index.eqn(255)
-					? idx === parsedPalletId
-					: meta.index.eqn(parsedPalletId)
-			);
-
-		return {
-			isValidPalletName,
-			isValidPalletIndex,
-			parsedPalletId,
-		};
-	}
-
-	/**
-	 * Identify if a pallet Identifier should be an index or a string. If it should
-	 * be an index return a number and if it should be a name return a string.
-	 *
-	 * @param palletId FRAME pallet identifier as a pallet name or index
-	 */
-	private static palletIdxOrName(palletId: string): string | number {
-		const maybeIdx = Number(palletId);
-
-		if (Number.isInteger(maybeIdx)) {
-			return maybeIdx;
-		}
-
-		return palletId;
 	}
 }
